refactor: move fmt.rs and io.rs under write_safe module
