+++ mm2src/common/write_safe/fmt.rs
