+++ mm2src/common/write_safe/mod.rs
@@ -1,0 +1,4 @@
+#[macro_use]
+pub mod fmt;
+#[macro_use]
+pub mod io;
