+++ mm2src/db_common/src/sql_create.rs
@@ -1,8 +1,8 @@
 use crate::sql_constraint::SqlConstraint;
 use crate::sql_value::{FromQuoted, SqlValue};
 use crate::sqlite::StringError;
-use common::fmt::{WriteSafe, WriteSafeJoin};
-use common::write_safe;
+use common::{write_safe,
+             write_safe::fmt::{WriteSafe, WriteSafeJoin}};
 use rusqlite::{Connection, Error as SqlError, Result as SqlResult};
 use std::fmt;
